Restructure As A Collection
Begin restructuring as a collection to modernize the resources in this role. Currently the filter_plugin used in the CLI installation is only available via configuration overrides. Moving it to a collection allow it to be used safely next to other ansible workloads. This also allows users to separate out their concerns like inventory into a repository that doesn't carry code. this enables users to adopt improvements in the roles and playbooks without having to spend time resolving merge conflicts with upstream.